Fix bug that was preventing creating new editions of transactions/places
The problem was that when you use the accessor for a hash or array field, mongoid puts it in the changed attributes. So when we came to copy an old edition it looked like it was changed and complained because published editions can't be changed.

That's an odd design decision from mongoid, but not one I feel brave enough to change at the moment. So for now we can get around that by using read_attribute