Refactor local services code. Add LGIL fallback.
Refactoring:- the data model for local transactions was confusing. This refactoring changes to use
the same domain language as the Local Government Standards (Service, Interaction, Authority). Furthermore,
the JSON api (consumed by frontend) was changed slightly to make the Interaction the primary data fetched
when a SNAC is provided in the request for a publication.

Fallback: if the 'apply' interaction is missing, then the 'find info' interaction is fetched instead.