Allow items to be progressed after fact check received
The post-fact-check buttons hadn't been updated since we switched to bootstrap so didn't work.
We also didn't have an integration test for this step. I added a failing test (and made a
couple of corresponding changes to the test helpers) and then updated the helper code to fix
the feature.