Add a test for skipping fact check
Conversations with James Adam revealed that we've not modelled the 'skip fact check' option clearly,
masking it as a 'fact check received' action with an appropriate message. We probably out to
make that behaviour more explicit as an action in the workflow. Before making that kind of change
it seemed only right to have an integration test for the current behaviour.