Add EditionObserver class
This class holds the evented business logic which was previously
within the Edition (and Workflow module) model(s). These evented
sequences are now contained within this class.

ActiveRecord::Observers only associates the observer class with
the model class of the same name. It does not apply any
inheritance lookups, which is why the list of Objects to be
observed has been made explicit.