Add tests to make Searchable mixins reproducible
The Searchable module is mixed into the Edition model using metaclasses
within Publisher. Where we're trying to produce indexable content for
solr, there are cases where GovSpeak is bleeding through to the index.

These tests are in place to make sure that the underlying calls adhere
to a reproducible contract.