Remove gds-api-adapters gem as it's no longer used
The use of gds-api-adapters in publisher ended when we brought the publisher and panopticon models together. I've removed it as a dependency to reduce load times.

I also reorganised the Gemfile a bit to group together (but not yet label) a few related dependencies. This is in preparation for a few more of them being removed over coming weeks.