Let the NoisyWorkflow email describe actions for itself
It's been bothering me for a while that there's a huge method in Action that is only used in one other place and is really much more about presentation logic than model/business logic. Adding NoisyWorkflow#describe_action will allow us to remove that from Action and put it much closer to where it's actually used.