Update button names for converters to be more specific
From a discussion with @alext it was made apparant that the names
of both of the buttons currently provided aren't that helpful as
it implies you're converting the Edition in place. In actual fact,
this is creating a new Edition of the specific kind.

This change updates the copy of the buttons.