Fix a bug viewing an edition with a newer sibling.
When viewing a published edition which had a newer, unpublished sibling,
Publisher would fall over with a NoMethodError. This was due to a
misnamed method call in the admin buttons.

Added a basic test to catch this particular bug: we could probably use
more comprehensive tests around the edition workflow to catch any other
bugs.