Stop the new list filter failing where partial name doesn't match scope name
We were using one list of edition states to whitelist list parameters, but that
doesn't match the list of templates we use and so was failing. We should
probably make the naming consistent but that could be quite disruptive at
a point when so much content production is taking place.