Check to avoid re-registering archived content.
This avoids the "resurrection bug" where archived content would get
re-registered in Panopticon and restored to "live" status, making it
viewable on the site and indexed in search. We'll have code in
Panopticon to mark all editions as archived when their corresponding
artefact is archived, but this provides an extra guard against
resurrection.