Fix links to private frontend
The version of Plek that was being used replaced dashes in hostnames,
but until recently this wasn't a problem because it was using a hard-coded
mapping. It broke because Publisher stopped using that. Plek was changed
a while ago to not replace dashes (0.1.26). Could have moved to that, but
this way it should give a smoother transition to Plek 0.5.0 and then 1.0.0
later on.