Benefits pages should have editable part titles
As part of the work to get Welsh content added to gov.uk, content
designers want the ability to edit the part titles for benefits pages so
they can provide a Welsh title rather than the rigid (currently
provided) set of titles.

This change (by proxy) allows editing of part slugs as well. However,
there's code currently existing which redirects non-existing part slugs
to the root page of an Edition.