Add panopticon_edit_url as a helper method
We had a method on the model to get a URL to edit the artefact. That didn't really feel
like a model concern and is only used in one controller so I've brought it in line
here and will remove it from the model.

I considered using a decorator to wrap the edition and provide the method but that
seemed like a lot of overhead for a single method. We should consider something like
that if attempting a broader refactor.