Break RootController#index into a set of smaller methods
These methods are far from ideal and mix side-effects (setting instance vars)
with return values in an awkward way. The intent of these changes is to begin
breaking the main method into smaller pieces to make it a little clearer how
it's constructed and how we might move to a cleaner implementation.