Correctly handle Panopticon registration in batch publish
This would break when an edition was published but
Panopticon timed out, as the edition could not be
published a second time to resend the notification.

We now fallback to invoking the registration method
again if the edition was published in a previous
attempt.

We think this only worked previously due to a bug
in the state_machine gem which was fixed in 1.2.0
(bumped with content models in previous commit).