Add autosave to editions form.
Autosave is performed only when the form becomes 'dirty', currently
dirty state is set by a keyup event on the form, the autosave check interval
runs approx every minute.
An autosave message is currently displayed when a successful save request completes,
this is also used by the autosave integration test.