Remove distinction between business/citizen in email alerts
This leaves a little bit of lingering debt as we still have separate email
groups. I've not addressed that here as I think we should talk a bit more
about how we want to handle the email alerts and then any changes will involve
co-ordinating changes to config as well as to this code.