Remove PublicationMetadata class as it added complexity
This one's a bit weird. It's included in publisher but referenced in govuk_content_models.
It dates from when there was a deeper separation between publisher and panopticon and
made a little more sense back then, but now it's a lot of code to extract a few fields
from a model.

There's a separate question of what we actually want to do with the metadata tab as it
doesn't feel quite right, but that's out of scope here.