[#51816839] Add shell out to nagios singal script
In order to provide a passive nagios check to signal the success of the
daily import of data I have added exception catching and signalling to
nagios through the local_transactions:fetch task. If there was an error
the message or class name will be passed through as the description.

The local_authority_import_check script is provided by the puppet
configuration such that all nagios configuration can be kept in one
place. The script takes a code (0 - success, 1 - error) and a
description of the status.

If an instance of publisher fails to lock then it will signal that it
was successful. That is to keep all the checks in sync and to remove
lingering errors from previous runs.

I have wrapped up the http fetching such that it throws an error for any
status but 200 such that we know whether the request failed through
nagios.