Add redirects for old /admin routes
It may be possible to do this in a single route but I wanted
to make it clear that we support both and avoid too much regexp
syntax.

There aren't any tests for this. It's pretty straightforward
rails code but I'd be happy to add a few if there are strong
feelings on the matter.