Make enhancements eager_load! friendly
Sidekiq calls eager_load! on a Rails app which
was loading the enhancements when they were under
app/models/, now that they are under lib/ it won't
eager load them.

The eager loading combined-with initializer
load_model_enhancements was causing files to
get required multiple times and the aliases
in User enhancements to be setup multiple
times which ended up causing a loop.