Add check for Return-Path header.
Since Return-Path is blank for Out of Office replies (checking factcheck
mail responses), this would be indistinguishable from the field not being
present if just using `mail[header].to_s`.

As well as checking the value of the header, I've added a check that
the header key returns a Mail::Field too. If the header is not present
NilClass is returned.