Configure the fact check script with a YAML file.
Instead of configuring the mail fetcher with a Ruby file that we dump
into the `config` directory at deploy time, we can parse a YAML file
with the settings for both the mail fetcher and the logic to generate
and parse fact check addresses. This should give us one single place to
put settings for both sending and receiving these emails.