Pull out shared code to set up state reporter.
Instead of having this declared twice in the Sidekiq worker and the
publication controller, let's set it up in an initialiser. We have to
set it up as a memoised accessor rather than assigning the value,
because it's hard to guarantee that the new initialiser will run after
the initialiser for the Statsd client. Memoising it means we can set up
test expectations without having to muck around with `any_instance`.