Convert activity forms to edition's attributes
www.agileplannerapp.com/boards/173808/cards/5629

Editors have reported issues with the existing
pattern of saving the edition form via AJAX before
doing a transition. Due to these multiple submissions
we suspect that on rare occations transition happens
but edition doesn't get saved.

Editors assume that when a transition happens, the
edition form changes are saved, and don't think of
it as a 2 step process.

This change ensures that transition is performed
using attributes within the edition form, which
ensures that we save the edition and perform the
transition in a single request.