Fix local-transaction validation error path.
There's a difference in the way template paths are resolved between
doing a `render :action` and `render :template`.  The latter was finding
the spurious local_transactions show template that had been left behind
from the refactoring in #256.

This also cleans up the redundant template (it's not used, and can't
work any more as it references partials that no longer exist.)