Send published documents to the content store.
Does not include the full content of the documents
as this is only for the purpose of emailing topic
updates.

I've made an assumption here that editions where
fact check was skipped should be considered minor
updates.

Future changes as part of this work:
- Adding topic/browse page associations

Future changes not part of this work:
- Migrating to full content store tag format
- Adding full body of content
- Fully implement content store functionality,
  removing "placeholder" status