Support sending republishes to content store
Update the `render_for_publishing_api` method to accept a parameter
telling it whether it's a republish event.  If so, it returns an
`update_type` field of `republish`.  This is understood by the content
store, and allows us to re-register content with the content store,
without causing actions like email alerts to happen.