Remove "alternate title" field from publisher
This removes the functionality to create and edit alternate titles for
editions within publisher. The intention is to get this change
deployed first before making an export of the data before it's finally
removed from the respective dependencies (govuk_content_models) and
database.

Story: https://www.agileplannerapp.com/boards/173808/cards/8145