Reintroduce previous important note workflow
Previously users could easily append to existing important notes. When
updating the important note model it wasn’t clear whether this was a
user need or an artefact of the implementation. It is a need — users
are often appending information to the end of an existing note (eg 2i
user and schedule info)

* Prepopulate important note field with existing note
* If the user blanks the field resolve the note (be consistent with old
behaviour)