Scroll to bottom before clicking "Add new part" in integration tests
Publisher has a set of common workflow buttons that sit within a bar
that's fixed to the bottom of the browser's window. This bar gets in
the way of Poltergeist "seeing" elements we want to click on.

To get around this there are three options:

1. Scroll to the bottom of the page and then click the link
2. Find the element on the page and trigger the "click" event
3. Restyle the page so that it doesn't have a fixed bar

For the purposes of this story, option 3 isn't realistic as it's going
to take a little more time and will require more conversations. Option
2 reads cleaner in terms of code, but takes longer to run on
average (tens of seconds longer) compared to option 1 where scrolling
to the bottom of the page is a fast action for WebKit.

This change goes with option 1 but tries to minimise the exposure of
this to a single helper method so that it can be refactored out later.

Please don't judge me too much for this :see_no_evil: