Avoid submitting business support form with JS
The update action for an edition relies on the value of the commit
button submitted with the form. By intercepting the form submit and
resubmitting after a DOM modification we lose that context.

* Don’t prevent default and avoid submitting form with JS
* Instead let the event bubble up and perform its default action,
submitting the form as the user originally intended

(This only seemed to be a problem within Firefox)