Bump api-adapters and content_models
This bumps gds-api-adapters and govuk_content_models. The change to
api-adapters includes support for the fields `public_timestamp` and 
`latest_change_note` in the Panopticon registerer. 

The change to govuk_content_models adds new methods to the `Edition`
model to access the major changes for an edition, the latest change note
for a major change, and the updated_at timestamp of the last major
change.