Migration to add 'the' to will_continue_on text
www.agileplannerapp.com/boards/173808/cards/8718

Frontend will stop adding 'the' to will_continue_on text
on business support editions:

https://github.com/alphagov/frontend/pull/695

This change allows editors to add 'the' when needed in
the best way possible, and keeps it consistent with
other formats.

This migration fixes existing data to contain that word.
Also, updated the hint text to suggest that editors need
to add 'the' to will_continue_on in business support.