Add endpoint for registering topic editions.
In order to get tag information for published
editions into Rummager (and thence to the services
and information page, among others) when a draft
topic is published, Panopticon needs to notify
Publisher so it can resubmit all published editions
tagged with that topic.

All this code can go away when drafts are handled
through the content store pipeline.