Treat first editions as 'major' changes
This changes the PublishedEditionPresenter so that, when presenting the 
first edition in a series, it always sets the 'update_type' field to
'major' change. 

Previously, first editions would never be sent as 'major' updates, as 
the UI is not shown to users. However, we always want to send email 
alerts for first editions of content, and display it in the latest 
changes feed for a sub-topic.