Add Rake task to run the processor
This adds a Rake task which, given a path to an existing CSV file, will 
parse the provided file, create an instance of the Processor object, 
and run the processor.

The explicit Logger instance is required in order for logging to be 
printed to the terminal without the need to run Rake with `--debug`.