topic_changes:prepare takes output file parameter
We want to be able to run the topic_changes:prepare task from rake.
To do this effectively, we need to be able to determine where the output
file is going to be generated.

This PR adds that as a required argument.  For safety, it also checks
that the output file doesn't exist already, to avoid overwriting files
by mistake.