Don't let 'Claim 2i' work when already claimed
If a user opens the “In review” list and leaves it for a while, they
may come back to it and click “Claim 2i” later, eg if they’ve finished
reviewing something else.

If in the meantime someone had already claimed the edition, the old
behaviour would overwrite that reviewer. The old reviewer wouldn’t be
notified, and the new reviewer wouldn’t know to tell them.

* Prevent claiming of 2i when there’s already a reviewer