Migration to set major_change for existing editions.
false is the default value for the fied on the model, however for all
the pre-existing editions, this will be set to nil (nor not even set).
This causes problems when archivind editions because the field gets
changed from nil to false, which trips up the validation that published
editions can't be changed. (This validation only allows the state and
updated_at field to be changed[1]).

This migration therefore sets all the existing editions to have a value
of false where it's not currently set.

[1] https://github.com/alphagov/govuk_content_models/blob/v26.2.0/app/models/edition.rb#L285-L297