Hide validation errors on newly added part
When a part gets added, it appears with error messages about missing
titles and slugs. These aren’t appropriate yet, and should be removed.
They will still show again should validation failed.

Using client side to handle this as on the server side it’s difficult
to differentiate between a just added part and a just added part that
has not passed validation checks on save.