Prevent intermittent failing video/campaign tests
Tests using javascript which submit a form that cannot be saved with
ajax (eg campaign image upload or video caption upload) were
intermittently failing. Within save_edition there was a check for a
dynamic saving message. Whilst this does show, it is likely that
occasionally the form would submit and the page would change before
this message appeared, leading to a failing test.

Instead pull the post-save assertions out of the save_edition method
into a separate one that takes into account whether ajax is being used.
Then in the video and campaign tests explicitly assert that ajax isn't
used.