Make multi-parameter date handling timezone aware
https://www.pivotaltracker.com/story/show/91500052

content designers noticed that scheduling a edition 00:01 was
showing-up as 01:01 in the publishing tool.

when the edition schedule time is submitted through the form, the
`ApplicationController#squash_multiparameter_datetime_attributes`
method squashes multi-parameter datetime into a single attribute
before saving the field in the database.

this logic was ignoring the local timezone, and made all field
handling logic beyond this point think that `publish_at` field
is in UTC.

using `Time.zone.local` preserves the timezone information, and
fixes this issue.

also correcting the integration test that was masking this issue.