Fix static file serving to work with nginx
We would like to offload (static file) CSV reports to be served by nginx
rather than rails. [This is possible](http://wiki.nginx.org/XSendfile) by setting
the correct header in the rails config.

The header was already set but it was configured to work with Apache, which we
no longer use.