Update all non-archived editions
If a document has no published edition, or the published edition was
already tagged with the required page, editions in draft state were not
also being updated.

Fix this by fetching all non-archived editions, instead of just the
published edition.

Change the output messages to start with what action or error happened,
so that it's easy to scan down the output for slugs which failed.

Also, refactor to put the code which adds the tag into a private method
instead of being inline, and to pass editions to the `import_error`
reporting method so that both the slug and version number can be
reported.