Republish after changing mainstream browse pages
After content has been tagged with mainstream browse pages, we need to
republish it to the publishing-api and to panopticon, so that search and
other information is updated correctly.

We republish whether we've made changes or not in this run, since a
previous run might have made the changes to the publisher database, but
then received errors when publishing.  Simply republishing always means
we can handle such errors cleanly.