Support ordering browse_page tags when importing
Add a third column to the CSV being used to import mainstream browse
tags to indicate whether the tag should be applied as a "primary" tag
(ie, at the start of the list of browse pages, so used for breadcrumbs
on the page, or at the end of the list).

Complain if we get any value other than "TRUE", "FALSE" or unspecified
from this field, so we don't get confused when converting from the value
in the CSV to a boolean.