Arrange mainstream browse pages in correct order
The list of tags is not orderable, but is shown during editing in the order that the tags were added. However, when the draft is saved and then re-displayed, the tags appear in alphabetical order. This has caused some confusion in the past amongst editors.

Similar solution to: https://github.com/alphagov/policy-publisher/pull/66