Make Fact Check email fetching more robust
This is an attempt to make the fact check email fetcher more robust and
more transparent. It seems like errors extending the lock time in Redis
may be causing emails to not be processed correctly and exit the loop
early, so we now rescue from the after_message hook and report errors
to Errbit.

Also added a few extra instances of Errbit logging in order to find
other errors in production.