Persist each Area's slug and GSS code side-by-side
We're moving away from the use of slugified area names in
Publisher/BusinessSupport.

Until we're ready to swap completely to the use of GSS codes, the publishing
tools need to support both types of identifier.

The solution I've chosen is unarguably gross. (I've combined the slug and the
GSS code in value fields in the editor's form.) There are definitely other
approaches we could take, but the main one I've decided against is making more
requests to the Imminence API during an update of a BusinessSupportEdition.

(If we were to move updating Editions to a worker, I'd feel better about
depending on the availability and responsiveness of external services there. As
it stands, I don't think there are any plans for that to happen. It could also
be that this isn't really a concern, and we could just make those requests..?)

I've made the change to the area identifier format directly in the areas.js.erb
template. I'm open to suggestions of where that might be better placed.
Possibly an AreasPresenter?