Store Areas metadata in their <option> elements
This isn't an essential change, but I hope putting the Areas data on the page
only once will help make this code easier to understand.

Until now the Business Support Scheme's selected and available-to-select Areas
were being rendered both as markup in the <option> elements and as a JavaScript
object for use by the "All of the UK" and "All of England" magic checkboxes.

If it's potential ineffiency offends anybody, I'd be happy for the `areas`
function to cache its data once processed, but assumed we'd prefer the
readability of this approach for now.