Fix bug in republisher scheduler
`edition.id` returns a `BSON::ObjectId` object and not a string.

When Sidekiq gets this object as an argument it calls `.to_json` on it
and that will crash the `PublishingAPINotifier` because it expects a
string with the ID. `BSON::ObjectId#to_s` returns the ID we want.

Commit includes a simple test to verify that the
`PublishingAPINotifier` can handle whatever arguments we're sending it.