Persist each Area's slug and GSS code side-by-side
We're moving away from the use of slugified area names in Publisher and the
Business Support apps.

Until we're ready to swap completely to the use of GSS codes, the publishing
tools need to support both types of identifier.

The solution I've chosen is unarguably gross. (I've combined the slug and the
GSS code in value fields in the editor's form.) There are definitely other
approaches we could take, but the main one I've decided against is making more
requests to the Imminence API during an update of a BusinessSupportEdition.