Convert parted editions
This tests for parted editions conversions such as programme editions
and guide editions.

When converting from a non-parted to a parted edition, the whole_body
of the source will be copied to a new part of the target class.

If the target class also has default parts, they will be created in addition to
the part receiving the whole_body.

The code to make this test pass is in govuk_content_models.