Fix sorting of GSS codes for regions for business support editions
Imminence still returns a fake 'England' region, which doesn't have a GSS
code. That means that there will be a nil in an array of GSS codes for all
areas returned from Imminence, which needs to be removed by `compact` before
the array can be sorted; otherwise we get `ArgumentError: comparison of
NilClass with String failed` when visiting the edit page for any business
support edition.