Exclude areas without GSS code from Area class methods
Areas are only used for business support editions, and we can only use areas
there which have GSS codes. At the moment Imminence adds a fake England region
into the response from MapIt and that area doesn't have a GSS code, so we
need to exclude it from the areas presented to users in Publisher. We should
in general handle missing data from other applications gracefully, so this
change will prevent Publisher erroring on any areas which don't have a GSS
code.