Add rake task for detecting "ghost" interactions
The CSV of interactions from local.directgov includes a complete list of all interactions in their database.  Those that have been deleted are marked with an "X" instead of a URL.  Our importer handles this to add/remove/update our DB so it mirrors this file.  However, we've noticed lots of "ghost" interactions where we have a record of an interaction that does not appear in the CSV file.  This shouldn't happen, but it does.  While we get to the root of the problem this rake task allows us to generate a CSV of all interactions that are in our DB but not in the CSV.

We iterate over all the local interactions stored against all the local authorities in our DB and so-doing we can detect 3 types of ghost:

 * interaction_in_input_to_be_deleted - any interaction that still lives in our DB but is listed as an "X" in the CSV.  Assumption here is that we simply haven't run the importer yet and it will be cleared on next import.
 * interaction_not_in_input - an interaction in our DB where the authority exists in the CSV but the interaction does not
 * authority_not_in_input - an interaction in our DB where the authority does not exist in the CSV