Fetch local.directgov data directly when checking for ghosts
We assumed that the file would be on disk from the import that we'd run
before, but that actually populates a tmp file rather than writing to
disk more pemanently, so we can't rely on it.

We also put the status of the interaction into the output file as it
might be useful.