Add check_for_ghosts:destroy task
Instead of just reporting all the ghost urls this task will actually
destroy the interaction if it is a ghost.  By default it only destroys
ghosts with the status :interaction_not_in_input, but you can pass
arguments to the task telling it which statuses to remove.

The rationale here is that :authority_not_in_input might be an authority
whose urls we manage from elsewhere (e.g. the new 2015 NI councils) and
:interaction_in_input_to_be_deleted should be removed by the next import
we run.