Change how we destroy ghost interactions
When actually running the destroy script we noticed that it would report an inconsistent number of interactions and deletions.  Cross-referencing with the detection script revealed that not all interactions were being iterated over and so not all ghosts were being destroyed in a single pass.

Our first thought was that iterating over the mongo cursor and modifying the result-set by deleting some of the results might be an issue.  We changed from `local_authority.local_interactions.each` (iterating over a cursor) to `local_authority.local_interactions.to_a.each` (flattening the cursor to an array and iterating over that).  This didn't help (but we've left it in just in case).

Our next thought was to change from `lai.destroy` (destroying the object directly) to `la.local_interactions.delete(lai)` (removing the object from it's container and destroying it implicitly).  Again this didn't solve the issue.

Our final solution is the one in this commit.  We don't destroy any objects directly, we collect all the non-ghost interactions and once we've seen everything for a given authority we use `la.local_interactions.substitute` to set the array to just those interactions.  This seems to do the right thing.  We think that because the interactions are embedded in the authority asking mongo to perform lots of actions on the authority in quick succession means things got "clogged up" and some operations failed.  We're not entirely sure though as we're on an old version of mongo and mongoid and it looks like there are plenty of reports of removing embedded documents simply not working.