Output "destroyed" ghosts after we destroy them
If the call to `substitute` fails we end up with the output.csv containing interactions we haven't actually destroyed yet.  To avoid this we collect the interactions to remove during the main loop and move writing them to the output after the call to `substitute` in the `on_authority_change` block.