Don't destroy when directgov file is shorter than expected
Running the destroy task when the directgov file is empty (or significantly shorter than we expect) would remove more data than we expect.  We put in place a limit where the task won't do anything unless the file has at that number of rows or more.  If this is not the case it will halt.  The limit can be specified in an ENV var to allow the task to proceed should the file start getting smaller.  We can then issue a PR to reduce the limit.  Note that the count here is not the length of the file.  We ignore rows with no SNAC code, but do count those with 'X's.

The limit is set at 100,000 for the moment, as we expect about ~102,000 interactions to be present in the file (including 'X's, excluding rows without a SNAC code) resulting in ~100,600 "live" interactions.