Discard publishing api draft when artefact is destroyed...
In the event of an artefact being destroyed the publishing-api needs to
be notified to discard the draft content item as the artefact slug may
be used again, eg. to draft a different type of edition.
The artefact would have a different content_id from the publishing api
unless the draft item is discarded.