Use strong parameters.
All the format types are subclasses of Edition, and share some core fields but
then have very differing sets of extra fields; we ensure that each subtype
only permits the parameters it is expecting.

The available fields depend on the edition subclass. In the case of an
update, we already have that as @resource, but on create we get it from
the :kind param. Turning a param into a constant is not really safe, so
we do it the other way round and turn the resource class into a symbol.