Don't fake filesystem in reporter tests.
FakeFS now causes issues with autoloaded code in tests, as any code that has
not already been loaded in that test run can no longer be found. Rather than
copy the code of the whole application and all its gems into the fake FS, it
is simpler to just stub out the paths to use temp dirs.

This involved a minor change to the report generator itself to return the path
from a method that can be stubbed, rather than a constant.