Make Publisher use Publishing API V2
This commit makes this app use the Publishing API V2. There should be
no changes in functionality.

I've refactored the some of the code dealing with `update_type`.
Because the signature of the sidekiq worker changes, we have to keep
supporting the old argument type. Otherwise jobs enqueued on
deploy-time with `normal` as the second argument would be
misinterpreted. This can be removed after deploy.