Add start button text input on  Simple Smart Answer form page
As requested in this trello card https://trello.com/c/CUe4odRB/45-make-start-now-editable-in-simple-smart-answers
the need to add the capability to change the text of the start button for simple smart answers.

Here, I have added a new form attribute "start_button_text" with a default text of "Start Now"

Also included are integration tests that ensure that this start button text is rendered on the SimpleSmartAnswerEdition form page.