Remove rows from local_services.csv that we don't use
As part of the work to replace local directgov we need to understand which
local services are used by GOV.UK on local transaction pages so that the
new local links manager can prioritize these services over others than we
don't use.

The best way to do this would be to ask content api for a list of all local
transactions and to pluck the lgsl codes used from there.  Unfortunately
content api does not provide this so the easiest way to do it for now is to
ingest the local_services.csv stored here in publisher.  However, this csv
contains services that we *might* use, rather than the ones that we do
use, so we remove all the ones that aren't in use.

We got the list of used services as follows:

    LocalTransactionEdition.all.map(:&lgsl_code).uniq.sort

We then removed all the rows from local_services.csv that were not in this
list.

Note that we have one LocalTransactionEdition with lgsl_code 1473 which is
not in the csv at all.  Turns out it was entered incorrectly (it should be
1743) and fixed in 5411e984203fe57149089f06961ca76fe201db46.  Our next
commit will include a script to remove the old LocalService instances
not in this CSV and update the lgsl_code on the edition.